refactor: 프로젝트 및 사용자 도메인 기능 리팩토링
1. 패키지 구조 변경
2. 로그 설정
3. 사용자 도메인 조회 기능 요청 파라미터 수정
