+++ pom.xml
@@ -163,11 +163,58 @@
 		  <version>5.0.1</version>
 		</dependency>
 		
-		<!-- Spring Security -->
+		<!-- Spring Security Core -->
 		<dependency>
 		  <groupId>org.springframework.security</groupId>
 		  <artifactId>spring-security-core</artifactId>
 		  <version>5.7.3</version>
+		</dependency>
+		
+		<!-- Spring Security Web -->
+		<dependency>
+		  <groupId>org.springframework.security</groupId>
+		  <artifactId>spring-security-web</artifactId>
+		  <version>5.7.9</version>
+		</dependency>
+		
+		<!-- Spring Security Config -->
+		<dependency>
+		  <groupId>org.springframework.security</groupId>
+		  <artifactId>spring-security-config</artifactId>
+		  <version>5.7.9</version>
+		</dependency>
+		
+		<dependency>
+		  <groupId>org.slf4j</groupId>
+		  <artifactId>slf4j-api</artifactId>
+		  <version>1.7.36</version>
+		</dependency>
+		
+		<!-- SLF4J Log4j2 Connection -->
+	    <dependency>
+	      <groupId>org.apache.logging.log4j</groupId>
+	      <artifactId>log4j-slf4j-impl</artifactId>
+	      <version>2.13.3</version> <!-- 최신 버전 확인 후 사용 -->
+	    </dependency>
+	
+	    <!-- Log4j2 -->
+	    <dependency>
+	      <groupId>org.apache.logging.log4j</groupId>
+	      <artifactId>log4j-api</artifactId>
+	      <version>2.13.3</version>
+	    </dependency>
+	
+	    <dependency>
+	      <groupId>org.apache.logging.log4j</groupId>
+	      <artifactId>log4j-core</artifactId>
+	      <version>2.13.3</version>
+	    </dependency>
+	    
+	    <!-- JSON -->
+	    <dependency>
+		  <groupId>com.fasterxml.jackson.core</groupId>
+		  <artifactId>jackson-databind</artifactId>
+		  <version>2.13.0</version>
 		</dependency>
 	</dependencies>
