+++ src/main/java/kr/co/logitics_track_360/user/dto/JoinRequestDto.java
@@ -1,4 +1,4 @@
-package user.dto;
+package kr.co.logitics_track_360.user.dto;
 
 public class JoinRequestDto {
 	private String userId;
