+++ src/main/java/kr/co/logitics_track_360/user/dto/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package user.dto;
+package kr.co.logitics_track_360.user.dto;
 
 public class LoginResponseDto {
     private String userId;
