+++ src/main/resources/db/migration/user/V1__create_table.sql
@@ -1,4 +1,4 @@
--- CREATE TABLE
+-- Create Table
 CREATE TABLE TBL_USER (
     USER_ID VARCHAR2(50) NOT NULL PRIMARY KEY,
     PASSWORD VARCHAR2(100) NOT NULL,
@@ -16,7 +16,7 @@ ALTER TABLE TBL_USER ADD CONSTRAINT constraint_user_role CHECK (ROLE IN ('CUSTOM
 -- Unique
 CREATE UNIQUE INDEX unique_user_email ON TBL_USER (EMAIL);
 
--- TRIGGER
+-- Trigger
 CREATE OR REPLACE TRIGGER trigger_user_updated_at
 BEFORE UPDATE ON TBL_USER
 FOR EACH ROW
