+++ src/main/resources/egovframework/sqlmap/mappers/UserMapper.xml
@@ -3,41 +3,41 @@
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="user.dao.UserMapper">
-  <insert id="insertUser" parameterType="user.dto.JoinRequestDto">
+<mapper namespace="kr.co.logitics_track_360.user.dao.UserMapper">
+  <insert id="insertUser" parameterType="kr.co.logitics_track_360.user.dto.JoinRequestDto">
     INSERT INTO TBL_USER (
-        user_id, 
-        password, 
-        nickname,
-        role,
-        email, 
-        created_at
+        USER_ID, 
+        PASSWORD, 
+        NAME,
+        ROLE,
+        EMAIL,
+        PHONE,
+        CREATED_AT
     ) VALUES (
         #{userId}, 
         #{password}, 
-        #{nickname},
+        #{name},
         #{role},
-        #{email}, 
+        #{email},
+        #{phone},
         CURRENT_TIMESTAMP
-    );
+    )
   </insert>
 
-  <select id="selectBy" resultType="user.vo.UserVO">
-    SELECT user_id as userId, 
-           password, 
-           nickname,
-           role,
-           email
+  <select id="selectBy" parameterType="map" resultType="kr.co.logitics_track_360.user.vo.UserVO">
+    SELECT USER_ID as userId, 
+           PASSWORD, 
+           NAME,
+           ROLE,
+           EMAIL 
     FROM TBL_USER
       <where>
         <if test="userId != null">
-          AND user_id = #{userId}
+          AND USER_ID = #{userId}
         </if>
         <if test="email != null">
-          AND email = #{email}
+          AND EMAIL = #{email}
         </if>
     </where>
   </select>
-  
-  
 </mapper>
