+++ src/main/webapp/WEB-INF/config/egovframework/security/spring-security.xml
@@ -1,0 +1,47 @@
+<beans:beans xmlns="http://www.springframework.org/schema/security"
+             xmlns:beans="http://www.springframework.org/schema/beans"
+             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+             xsi:schemaLocation="
+                http://www.springframework.org/schema/security
+                http://www.springframework.org/schema/security/spring-security.xsd
+                http://www.springframework.org/schema/beans
+                http://www.springframework.org/schema/beans/spring-beans.xsd">
+
+    <http auto-config="true" use-expressions="true">
+     	<csrf disabled="true"/>
+     	<intercept-url pattern="/user/login.do" access="permitAll" />
+    	<intercept-url pattern="/user/join.do" access="permitAll" />
+        <intercept-url pattern="/admin/**" access="hasRole('ADMIN')" />
+        <intercept-url pattern="/user/**" access="hasRole('CUSTOMER')" />
+        <intercept-url pattern="/**" access="permitAll" />
+
+        <form-login
+            login-page="/login.do"
+            default-target-url="/main.do"
+            authentication-failure-url="/login.do?error=true"
+            username-parameter="username"
+            password-parameter="password" />
+
+        <logout logout-url="/logout.do"
+                logout-success-url="/login.do"
+                invalidate-session="true" />
+    </http>
+
+    <authentication-manager>
+        <authentication-provider>
+            <jdbc-user-service data-source-ref="dataSource"
+                users-by-username-query="
+                    SELECT USER_ID  AS username, 
+                           PASSWORD AS password, 
+                           1        AS enabled
+                    FROM TBL_USER
+                    WHERE USER_ID = ?"
+                authorities-by-username-query="
+                    SELECT USER_ID AS username, 
+                           ROLE    AS authority
+                    FROM TBL_USER
+                    WHERE USER_ID = ?" />
+        </authentication-provider>
+    </authentication-manager>
+
+</beans:beans>
