+++ src/main/webapp/WEB-INF/config/egovframework/springmvc/dispatcher-servlet/dispatcher-servlet.xml
@@ -10,8 +10,7 @@
            http://www.springframework.org/schema/context/spring-context.xsd
            http://www.springframework.org/schema/mvc
            http://www.springframework.org/schema/mvc/spring-mvc.xsd">
-  <context:component-scan base-package="web" />
-  <context:component-scan base-package="service.impl" />
+  <context:component-scan base-package="kr.co.logitics_track_360" />
   
   <mvc:annotation-driven />
