+++ src/main/webapp/WEB-INF/web.xml
@@ -13,6 +13,14 @@
       classpath:context-datasource.xml
     </param-value>
   </context-param>
+  
+  <context-param>
+    <param-name>contextConfigLocation</param-name>
+    <param-value>
+      classpath:context-datasource.xml
+      /WEB-INF/config/egovframework/security/spring-security.xml
+    </param-value>
+  </context-param>
 
   <!-- Spring Context -->
   <listener>
@@ -34,4 +42,15 @@
     <servlet-name>dispatcher</servlet-name>
     <url-pattern>*.do</url-pattern>
   </servlet-mapping>
+  
+  <!-- Spring Security -->
+  <filter>
+    <filter-name>springSecurityFilterChain</filter-name>
+    <filter-class>org.springframework.web.filter.DelegatingFilterProxy</filter-class>
+  </filter>
+	
+  <filter-mapping>
+	<filter-name>springSecurityFilterChain</filter-name>
+    <url-pattern>/*</url-pattern>
+  </filter-mapping>
 </web-app>
